requirements.txt
line-bot-sdk==2.0.1
Flask==2.0.1
requests==2.21.0
firebase-admin==4.4.0